Refactor #324: Moved authentication rate limiter into a dedicated middleware/authLimiter.js module
