+++ backend/server.js
@@ -10,7 +10,7 @@ const rateLimit = require("express-rate-limit");
 const helmet = require("helmet");
 
 const routes = require("./routes/index")
-
+const authLimiter = require("./middleware/authLimiter");
 // load environment
 dotenv.config();
 
@@ -182,23 +182,6 @@ if (process.env.NODE_ENV !== "production") {
     next();
   });
 }
-
-// auth limiter
-const authLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-
-  max: 20,
-
-  standardHeaders: true,
-
-  legacyHeaders: false,
-
-  message: {
-    success: false,
-
-    message: "Too many requests. Please try again later.",
-  },
-});
 
 // api limiter
 const apiLimiter = rateLimit({
